backend/pyproject.toml
[project]
name = "smart-manufact"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "docling>=2.69.1",
    "fastapi[standard]>=0.128.2",
    "httpx>=0.28.1",
    "minio>=7.2.20",
    "neo4j>=6.1.0",
    "numpy>=2.4.1",
    "openai>=1.109.1",
    "opensearch-py>=3.1.0",
    "pillow<12",
    "psycopg2-binary>=2.9.11",
    "pymupdf>=1.26.7",
    "python-multipart>=0.0.22",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "tqdm>=4.67.1",
    "uvicorn>=0.40.0",
]
